Refactor game data into modules
