+++ js/constants.js
@@ -1,0 +1,25 @@
+export const GRID_WIDTH = 12;
+export const GRID_HEIGHT = 8;
+export const STARTING_CASH = 150000;
+export const DAY_TICKS = 60;
+export const CANVAS_CELL = 56;
+export const AXIS_LETTERS = "ABCDEFGHIJKLMNOPQRSTUVWXYZ";
+export const SAVE_VERSION = 1;
+export const SAVE_SLOT_COUNT = 4;
+export const SAVE_STORAGE_KEY_PREFIX = "pulse-point-hospital-save";
+export const VIEW_MODE_STORAGE_KEY = "pulse-point-hospital-view-mode";
+export const ISO_TILE_WIDTH = 64;
+export const ISO_TILE_HEIGHT = 36;
+export const ISO_WALL_HEIGHT = 28;
+export const NEW_ROOM_HIGHLIGHT_TICKS = 36;
+export const MIN_SHOWCASE_ZOOM = 0.65;
+export const MAX_SHOWCASE_ZOOM = 1.8;
+export const SHOWCASE_PAN_PADDING = 0.35;
+export const AUDIO_SETTINGS_STORAGE_KEY = "pulse-point-hospital-audio-settings";
+export const DEFAULT_AUDIO_SETTINGS = { music: true, sfx: true };
+export const BACKGROUND_MUSIC_DURATION = 8;
+export const CORRIDOR_Y = GRID_HEIGHT - 0.25;
+export const ENTRANCE_X = 1.5;
+export const AGENT_PATIENT_SPEED = 0.18;
+export const AGENT_STAFF_SPEED = 0.22;
+export const TREATMENT_DURATION_TICKS = 18;
