+++ js/data/roomData.js
@@ -1,0 +1,732 @@
+export const PROPERTY_PARCELS = [
+  {
+    id: "atrium",
+    name: "Founders Atrium",
+    cost: 0,
+    description: "Starter plaza with enough space for a GP hub and diagnostics.",
+    x: 0,
+    y: 0,
+    width: 6,
+    height: 4,
+  },
+  {
+    id: "research",
+    name: "Innovation Lot",
+    cost: 48000,
+    description: "Unlocks the north labs and plenty of room for specialty clinics.",
+    x: 6,
+    y: 0,
+    width: 6,
+    height: 4,
+  },
+  {
+    id: "gardens",
+    name: "Wellness Gardens",
+    cost: 36000,
+    description: "Adds calming grounds for wards, psychiatry, and guest services.",
+    x: 0,
+    y: 4,
+    width: 6,
+    height: 4,
+  },
+  {
+    id: "emergency",
+    name: "Emergency Concourse",
+    cost: 52000,
+    description: "Extends the trauma wing for emergency, surgery, and imaging suites.",
+    x: 6,
+    y: 4,
+    width: 6,
+    height: 4,
+  },
+];
+
+export const ROOM_SIZE_LIBRARY = [
+  {
+    id: "compact",
+    label: "Compact (1x1)",
+    width: 1,
+    height: 1,
+    costMultiplier: 1,
+    upkeepMultiplier: 1,
+    severityBonus: 0,
+    environmentBonus: 0.2,
+    description: "Entry layout for a quick setup.",
+  },
+  {
+    id: "standard",
+    label: "Standard (2x1)",
+    width: 2,
+    height: 1,
+    costMultiplier: 1.35,
+    upkeepMultiplier: 1.2,
+    severityBonus: 0.5,
+    environmentBonus: 0.6,
+    description: "Adds an extra bay for machines and visitors.",
+  },
+  {
+    id: "deluxe",
+    label: "Deluxe (2x2)",
+    width: 2,
+    height: 2,
+    costMultiplier: 1.8,
+    upkeepMultiplier: 1.35,
+    severityBonus: 1,
+    environmentBonus: 1,
+    description: "Premium suite with ample space for advanced care.",
+  },
+];
+
+export const ROOM_INTERIOR_LIBRARY = [
+  {
+    id: "modern",
+    label: "Modern Steel",
+    cost: 1200,
+    upkeep: 35,
+    palette: {
+      base: "#1f2937",
+      mid: "#334155",
+      accent: "#60a5fa",
+      decor: "#93c5fd",
+    },
+    modifiers: { efficiency: 2, morale: 1 },
+  },
+  {
+    id: "soothing",
+    label: "Soothing Pastels",
+    cost: 1500,
+    upkeep: 30,
+    palette: {
+      base: "#fdf2f8",
+      mid: "#fce7f3",
+      accent: "#f472b6",
+      decor: "#34d399",
+    },
+    modifiers: { welfare: 3, environment: 2 },
+  },
+  {
+    id: "vibrant",
+    label: "Vibrant Pop",
+    cost: 1000,
+    upkeep: 25,
+    palette: {
+      base: "#0f172a",
+      mid: "#1e293b",
+      accent: "#f97316",
+      decor: "#facc15",
+    },
+    modifiers: { reputation: 2, morale: 2 },
+  },
+  {
+    id: "nature",
+    label: "Nature Retreat",
+    cost: 1350,
+    upkeep: 28,
+    palette: {
+      base: "#f1f5f9",
+      mid: "#bbf7d0",
+      accent: "#22c55e",
+      decor: "#65a30d",
+    },
+    modifiers: { welfare: 2, morale: 1, environment: 3 },
+  },
+  {
+    id: "art-deco",
+    label: "Art Deco Luxe",
+    cost: 1650,
+    upkeep: 38,
+    palette: {
+      base: "#0f172a",
+      mid: "#1e293b",
+      accent: "#fbbf24",
+      decor: "#f472b6",
+    },
+    modifiers: { reputation: 3, morale: 2 },
+  },
+  {
+    id: "sterile",
+    label: "Sterile Quartz",
+    cost: 1400,
+    upkeep: 32,
+    palette: {
+      base: "#f8fafc",
+      mid: "#cbd5f5",
+      accent: "#38bdf8",
+      decor: "#94a3b8",
+    },
+    modifiers: { efficiency: 2, environment: 1 },
+  },
+  {
+    id: "sunset",
+    label: "Sunset Oasis",
+    cost: 1550,
+    upkeep: 34,
+    palette: {
+      base: "#fff7ed",
+      mid: "#fed7aa",
+      accent: "#fb7185",
+      decor: "#f97316",
+    },
+    modifiers: { welfare: 2, morale: 2 },
+  },
+  {
+    id: "zenith",
+    label: "Zenith Glasshouse",
+    cost: 1850,
+    upkeep: 38,
+    palette: {
+      base: "#0b1120",
+      mid: "#134e4a",
+      accent: "#2dd4bf",
+      decor: "#f59e0b",
+    },
+    modifiers: { environment: 3, reputation: 1, efficiency: 1 },
+  },
+  {
+    id: "heritage",
+    label: "Heritage Copper",
+    cost: 1700,
+    upkeep: 33,
+    palette: {
+      base: "#1c1917",
+      mid: "#b45309",
+      accent: "#f97316",
+      decor: "#fde68a",
+    },
+    modifiers: { morale: 2, reputation: 2 },
+  },
+  {
+    id: "aurora",
+    label: "Aurora Pulse",
+    cost: 1780,
+    upkeep: 37,
+    palette: {
+      base: "#111827",
+      mid: "#312e81",
+      accent: "#a855f7",
+      decor: "#38bdf8",
+    },
+    modifiers: { efficiency: 2, welfare: 1, reputation: 2 },
+  },
+];
+
+export const ROOM_MACHINE_LIBRARY = [
+  {
+    id: "standard-kit",
+    label: "Standard Treatment Kit",
+    cost: 0,
+    upkeep: 0,
+    severityBoost: 0,
+    welfareBoost: 0,
+    description: "Baseline tools for routine care.",
+    color: "#cbd5f5",
+  },
+  {
+    id: "advanced-diagnostics",
+    label: "Advanced Diagnostics Rig",
+    cost: 4200,
+    upkeep: 80,
+    severityBoost: 1,
+    welfareBoost: 0.5,
+    description: "High-resolution scanners for tricky ailments.",
+    color: "#38bdf8",
+  },
+  {
+    id: "precision-suite",
+    label: "Precision Therapy Suite",
+    cost: 7600,
+    upkeep: 120,
+    severityBoost: 2,
+    welfareBoost: 1,
+    description: "Specialist pods that unlock elite treatments.",
+    color: "#f97316",
+  },
+  {
+    id: "holistic-support",
+    label: "Holistic Support Capsule",
+    cost: 3500,
+    upkeep: 55,
+    severityBoost: 0.3,
+    welfareBoost: 1.5,
+    description: "Comfort tech that keeps patients calm and cooperative.",
+    color: "#34d399",
+  },
+];
+
+export const ROOM_DECOR_LIBRARY = [
+  {
+    id: "calming-plants",
+    label: "Calming Planters",
+    cost: 800,
+    upkeep: 10,
+    environment: 2,
+    welfare: 1,
+    morale: 1,
+    reputation: 0,
+    description: "Greenery that softens clinical edges.",
+    color: "#22c55e",
+    isPlant: true,
+  },
+  {
+    id: "art-installation",
+    label: "Art Installation",
+    cost: 1100,
+    upkeep: 14,
+    environment: 1,
+    welfare: 1,
+    morale: 0,
+    reputation: 1,
+    description: "Gallery-worthy pieces for a prestige boost.",
+    color: "#a855f7",
+  },
+  {
+    id: "luxury-seating",
+    label: "Luxury Seating",
+    cost: 1500,
+    upkeep: 18,
+    environment: 1,
+    welfare: 2,
+    morale: 2,
+    reputation: 1,
+    description: "Plush lounges that keep visitors smiling.",
+    color: "#facc15",
+  },
+  {
+    id: "healing-sculpture",
+    label: "Healing Sculpture",
+    cost: 950,
+    upkeep: 12,
+    environment: 1,
+    welfare: 1,
+    morale: 1,
+    reputation: 1,
+    description: "Centerpiece art celebrating recovery stories.",
+    color: "#0ea5e9",
+  },
+  {
+    id: "snack-vendor",
+    label: "Snack Machine",
+    cost: 650,
+    upkeep: 8,
+    environment: 0.2,
+    welfare: 1,
+    morale: 1,
+    reputation: 0,
+    description: "Dispenses quick bites that cheer up waiting patients (and create a little litter).",
+    color: "#fb7185",
+    amenityType: "snack",
+    amenityLabel: "Snacks",
+  },
+  {
+    id: "atm-kiosk",
+    label: "ATM Kiosk",
+    cost: 950,
+    upkeep: 9,
+    environment: 0.5,
+    welfare: 0,
+    morale: 0,
+    reputation: 1,
+    description: "Cash-point services that earn small fees and reassure guests.",
+    color: "#0ea5e9",
+    amenityType: "atm",
+    amenityLabel: "ATM",
+  },
+  {
+    id: "hydration-station",
+    label: "Hydration Station",
+    cost: 700,
+    upkeep: 7,
+    environment: 1,
+    welfare: 1,
+    morale: 0,
+    reputation: 0,
+    description: "Water coolers and cups that keep everyone refreshed.",
+    color: "#60a5fa",
+    amenityType: "hydration",
+    amenityLabel: "Hydration",
+  },
+  {
+    id: "hanging-garden",
+    label: "Hanging Garden",
+    cost: 900,
+    upkeep: 11,
+    environment: 2,
+    welfare: 1,
+    morale: 1,
+    reputation: 0,
+    description: "Vertical planters that need routine watering but elevate the ambience.",
+    color: "#4ade80",
+    isPlant: true,
+  },
+];
+
+export const roomCatalog = [
+  {
+    id: "reception",
+    name: "Reception",
+    cost: 4500,
+    upkeep: 150,
+    roleRequired: ["assistant"],
+    reputation: 1,
+    description: "Welcomes new arrivals and keeps the queue orderly.",
+  },
+  {
+    id: "triage",
+    name: "Triage Desk",
+    cost: 5200,
+    upkeep: 180,
+    roleRequired: ["assistant", "nurse"],
+    reputation: 2,
+    description: "Sorts arrivals and smooths the waiting line before treatment.",
+  },
+  {
+    id: "gp",
+    name: "Doctor's Office",
+    cost: 6000,
+    upkeep: 200,
+    roleRequired: ["doctor"],
+    reputation: 2,
+    description: "Diagnoses patient ailments and sets their treatment plan.",
+  },
+  {
+    id: "psychiatry",
+    name: "Psychiatry",
+    cost: 14000,
+    upkeep: 360,
+    roleRequired: ["doctor"],
+    reputation: 3,
+    description: "Provides therapy for mental health disorders and stress ailments.",
+  },
+  {
+    id: "pediatrics",
+    name: "Pediatrics",
+    cost: 13000,
+    upkeep: 340,
+    roleRequired: ["doctor", "nurse"],
+    reputation: 3,
+    description: "Bright ward for children with extra play space and comfort.",
+  },
+  {
+    id: "maternity",
+    name: "Maternity Suite",
+    cost: 17000,
+    upkeep: 420,
+    roleRequired: ["doctor", "midwife"],
+    reputation: 4,
+    description: "Delivers new arrivals with calming decor and specialised care.",
+  },
+  {
+    id: "pharmacy",
+    name: "Pharmacy",
+    cost: 8000,
+    upkeep: 250,
+    roleRequired: ["nurse"],
+    reputation: 3,
+    description: "Dispenses medicine for low severity conditions.",
+  },
+  {
+    id: "ward",
+    name: "Ward",
+    cost: 12000,
+    upkeep: 360,
+    roleRequired: ["doctor", "nurse"],
+    reputation: 4,
+    description: "Treats moderate conditions requiring overnight care.",
+  },
+  {
+    id: "diagnostics",
+    name: "Diagnostics Lab",
+    cost: 15000,
+    upkeep: 420,
+    roleRequired: ["doctor"],
+    reputation: 4,
+    description: "Runs advanced tests to speed up treatment chains.",
+  },
+  {
+    id: "cardiology",
+    name: "Cardiology",
+    cost: 20000,
+    upkeep: 520,
+    roleRequired: ["doctor", "nurse"],
+    reputation: 5,
+    description: "Monitors heart health and treats rhythm-related illnesses.",
+  },
+  {
+    id: "orthopedics",
+    name: "Orthopedics",
+    cost: 19000,
+    upkeep: 480,
+    roleRequired: ["doctor", "therapist"],
+    reputation: 4,
+    description: "Fixes broken bones and mobility issues with rehab stations.",
+  },
+  {
+    id: "neurology",
+    name: "Neurology",
+    cost: 21000,
+    upkeep: 540,
+    roleRequired: ["doctor", "technician"],
+    reputation: 5,
+    description: "Maps brain activity with scanners and quiet monitoring pods.",
+  },
+  {
+    id: "dermatology",
+    name: "Dermatology",
+    cost: 10000,
+    upkeep: 260,
+    roleRequired: ["doctor"],
+    reputation: 2,
+    description: "Treats skin conditions with UV bays and soothing creams.",
+  },
+  {
+    id: "gastro",
+    name: "Gastroenterology",
+    cost: 16000,
+    upkeep: 400,
+    roleRequired: ["doctor", "nurse"],
+    reputation: 4,
+    description: "Investigates digestive disorders with scopes and recovery chairs.",
+  },
+  {
+    id: "dialysis",
+    name: "Dialysis",
+    cost: 15000,
+    upkeep: 420,
+    roleRequired: ["nurse", "technician"],
+    reputation: 4,
+    description: "Filters blood with high-tech machines for chronic patients.",
+  },
+  {
+    id: "physiotherapy",
+    name: "Physiotherapy",
+    cost: 12500,
+    upkeep: 320,
+    roleRequired: ["nurse"],
+    reputation: 3,
+    description: "Rehabilitates injuries and mobility issues with guided exercise.",
+  },
+  {
+    id: "oncology",
+    name: "Oncology",
+    cost: 22000,
+    upkeep: 560,
+    roleRequired: ["doctor", "technician"],
+    reputation: 5,
+    description: "Provides chemo bays and support zones for cancer patients.",
+  },
+  {
+    id: "burn",
+    name: "Burn Unit",
+    cost: 23000,
+    upkeep: 600,
+    roleRequired: ["doctor", "nurse", "therapist"],
+    reputation: 5,
+    description: "Cares for severe burns with hydro tanks and pain therapy booths.",
+  },
+  {
+    id: "dental",
+    name: "Dental Surgery",
+    cost: 9000,
+    upkeep: 260,
+    roleRequired: ["dentist", "assistant"],
+    reputation: 3,
+    description: "Restores smiles with chrome chairs and high-powered polishers.",
+  },
+  {
+    id: "ophthalmology",
+    name: "Ophthalmology",
+    cost: 18000,
+    upkeep: 520,
+    roleRequired: ["doctor", "technician"],
+    reputation: 4,
+    description: "Performs delicate eye work with laser rigs and dark testing rooms.",
+  },
+  {
+    id: "surgery",
+    name: "Operating Theatre",
+    cost: 26000,
+    upkeep: 900,
+    roleRequired: ["surgeon", "nurse"],
+    reputation: 6,
+    description: "Performs complex procedures for severe conditions.",
+  },
+  {
+    id: "icu",
+    name: "Intensive Care Unit",
+    cost: 24000,
+    upkeep: 820,
+    roleRequired: ["doctor", "nurse"],
+    reputation: 5,
+    description: "Stabilises critical patients and boosts emergency reputation.",
+  },
+  {
+    id: "emergency",
+    name: "Emergency Department",
+    cost: 24000,
+    upkeep: 780,
+    roleRequired: ["doctor", "nurse", "security"],
+    reputation: 6,
+    description: "Handles urgent arrivals with crash bays and trauma carts.",
+  },
+  {
+    id: "research",
+    name: "Research Lab",
+    cost: 18000,
+    upkeep: 550,
+    roleRequired: ["researcher"],
+    reputation: 5,
+    description: "Unlocks new cures and room upgrades.",
+  },
+  {
+    id: "cafeteria",
+    name: "Wellness Cafeteria",
+    cost: 9000,
+    upkeep: 260,
+    roleRequired: ["assistant"],
+    reputation: 2,
+    description: "Keeps patients waiting happy and lifts staff morale.",
+  },
+  {
+    id: "gourmet",
+    name: "Gourmet Kitchen",
+    cost: 11000,
+    upkeep: 320,
+    roleRequired: ["chef", "assistant"],
+    reputation: 3,
+    description: "Serves signature meals to energise staff and pamper VIPs.",
+  },
+  {
+    id: "giftshop",
+    name: "Gift Shop",
+    cost: 6500,
+    upkeep: 160,
+    roleRequired: ["assistant"],
+    reputation: 2,
+    description: "Sells plush mascots and boosts visitor satisfaction.",
+  },
+  {
+    id: "clowncare",
+    name: "Clown Care Ward",
+    cost: 9000,
+    upkeep: 260,
+    roleRequired: ["entertainer"],
+    reputation: 3,
+    description: "Inspired by Theme Hospital antics to keep spirits high.",
+  },
+  {
+    id: "holotheatre",
+    name: "Holotheatre",
+    cost: 12500,
+    upkeep: 320,
+    roleRequired: ["entertainer", "assistant"],
+    reputation: 3,
+    description: "Projects Biing-style cabaret holograms minus the risqué bits.",
+  },
+  {
+    id: "zen",
+    name: "Zen Garden",
+    cost: 8000,
+    upkeep: 220,
+    roleRequired: ["therapist"],
+    reputation: 2,
+    description: "Restorative retreat with koi ponds for stressed staff.",
+  },
+  {
+    id: "securityoffice",
+    name: "Security Office",
+    cost: 7000,
+    upkeep: 200,
+    roleRequired: ["security"],
+    reputation: 2,
+    description: "Keeps troublemakers in check and smooths emergency response.",
+  },
+  {
+    id: "billing",
+    name: "Billing Office",
+    cost: 10500,
+    upkeep: 300,
+    roleRequired: ["accountant"],
+    reputation: 3,
+    description: "Optimises invoices for better payouts and financial tracking.",
+  },
+  {
+    id: "marketing",
+    name: "Public Relations Suite",
+    cost: 9500,
+    upkeep: 220,
+    roleRequired: ["marketer"],
+    reputation: 2,
+    description: "Launch campaigns to attract lucrative patient groups.",
+  },
+  {
+    id: "staffroom",
+    name: "Staff Lounge",
+    cost: 7000,
+    upkeep: 160,
+    roleRequired: [],
+    reputation: 1,
+    description: "Boosts staff morale and reduces burnout.",
+  },
+];
+
+export const baseSeverityMap = {
+  reception: 0,
+  triage: 1,
+  gp: 2,
+  psychiatry: 2,
+  pediatrics: 2,
+  maternity: 3,
+  pharmacy: 1,
+  ward: 2,
+  diagnostics: 2,
+  cardiology: 3,
+  orthopedics: 3,
+  neurology: 3,
+  dermatology: 2,
+  gastro: 3,
+  dialysis: 3,
+  physiotherapy: 2,
+  oncology: 4,
+  burn: 4,
+  dental: 2,
+  ophthalmology: 3,
+  surgery: 4,
+  icu: 4,
+  emergency: 3,
+  cafeteria: 0,
+  gourmet: 0,
+  giftshop: 0,
+  billing: 0,
+  marketing: 0,
+  clowncare: 1,
+  holotheatre: 1,
+  zen: 1,
+  research: 0,
+  securityoffice: 0,
+  staffroom: 0,
+  radiology: 3,
+};
+
+export const defaultDesignProfile = {
+  sizes: ROOM_SIZE_LIBRARY.map((option) => option.id),
+  interiors: ROOM_INTERIOR_LIBRARY.map((theme) => theme.id),
+  machines: ROOM_MACHINE_LIBRARY.map((machine) => machine.id),
+  decor: ROOM_DECOR_LIBRARY.map((decor) => decor.id),
+  defaultMachines: ["standard-kit"],
+};
+
+roomCatalog.forEach((room) => {
+  room.baseSeverity = baseSeverityMap[room.id] ?? 1;
+  room.designProfile = {
+    sizes: [...defaultDesignProfile.sizes],
+    interiors: [...defaultDesignProfile.interiors],
+    machines:
+      (baseSeverityMap[room.id] ?? 1) > 0 ? [...defaultDesignProfile.machines] : ["standard-kit"],
+    decor: [...defaultDesignProfile.decor],
+    defaultMachines: [...defaultDesignProfile.defaultMachines],
+  };
+  if ((baseSeverityMap[room.id] ?? 1) <= 0) {
+    room.designProfile.machines = ["standard-kit"];
+  }
+});
+
