+++ js/data/staffData.js
@@ -1,0 +1,99 @@
+export const staffCatalog = [
+  {
+    role: "doctor",
+    namePool: ["Dr. Nova Halley", "Dr. Elias Quill", "Dr. Miri Vega", "Dr. Jun Pierce"],
+    salary: [900, 1100],
+    traits: ["Bedside manner", "Diagnostic genius", "Fast learner", "Research enthusiast"],
+  },
+  {
+    role: "nurse",
+    namePool: ["Nurse Aiko", "Nurse Lucien", "Nurse Piper", "Nurse Saffron"],
+    salary: [650, 850],
+    traits: ["Efficient", "Empathetic", "Inventory hawk", "Inspiring"],
+  },
+  {
+    role: "assistant",
+    namePool: ["Maya Front", "Theo Ledger", "Rae Burst", "Ivy Kios"],
+    salary: [450, 600],
+    traits: ["Charismatic", "Queue master", "Patient whisperer", "Energetic"],
+  },
+  {
+    role: "janitor",
+    namePool: ["Vik Sweep", "Chaz Buff", "Vera Shine", "Kato Mop"],
+    salary: [380, 520],
+    traits: ["Meticulous", "Fast cleaner", "Mechanic", "Green thumb"],
+  },
+  {
+    role: "facility",
+    namePool: ["Mina Grove", "Odin Wells", "Clara Bloom", "Zed Rinse"],
+    salary: [420, 600],
+    traits: ["Plant whisperer", "Spill scout", "Sparkling tidy", "Logistics ace"],
+  },
+  {
+    role: "researcher",
+    namePool: ["Dr. Lyra Code", "Analyst Vex", "Dr. Reva Coil", "Milo Synth"],
+    salary: [700, 950],
+    traits: ["Innovator", "Grant magnet", "Focused", "Collaborative"],
+  },
+  {
+    role: "marketer",
+    namePool: ["Ada Pitch", "Rowan Buzz", "Nix Viral", "Tara Prism"],
+    salary: [500, 680],
+    traits: ["Storyteller", "Analytical", "Hype builder", "Networked"],
+  },
+  {
+    role: "surgeon",
+    namePool: ["Dr. Mira Scalpel", "Dr. Orion Bloom", "Dr. Zia Voss", "Dr. Helio Mars"],
+    salary: [1200, 1500],
+    traits: ["Calm hands", "Rapid recovery", "Team lead", "Minimal scarring"],
+  },
+  {
+    role: "technician",
+    namePool: ["Ray Pulse", "Iris Beam", "Nico Magnet", "Kara Scan"],
+    salary: [580, 760],
+    traits: ["Precision", "Calibrates fast", "Empathic", "Data driven"],
+  },
+  {
+    role: "accountant",
+    namePool: ["Pia Ledger", "Miles Tally", "June Audit", "Cai Balance"],
+    salary: [520, 720],
+    traits: ["Eagle eye", "Negotiator", "Budget hawk", "Calm"],
+  },
+  {
+    role: "midwife",
+    namePool: ["Mara Bloom", "Leah Cradle", "Ina Dove", "Ruth Lulla"],
+    salary: [700, 900],
+    traits: ["Gentle", "Quick hands", "Reassuring", "Songbird"],
+  },
+  {
+    role: "therapist",
+    namePool: ["Galen Flex", "Noor Mend", "Ravi Align", "Elara Sooth"],
+    salary: [680, 880],
+    traits: ["Motivator", "Massage pro", "Mindful", "Stretch guru"],
+  },
+  {
+    role: "entertainer",
+    namePool: ["Bobo Spark", "Luna Gag", "Pip Confetti", "Dex Pratfall"],
+    salary: [520, 720],
+    traits: ["Slapstick", "Improv", "Balloon artist", "Banters"],
+  },
+  {
+    role: "security",
+    namePool: ["Axel Guard", "Mira Watch", "Jett Shield", "Tova Brace"],
+    salary: [550, 750],
+    traits: ["Alert", "Crowd control", "Cool head", "Swift"],
+  },
+  {
+    role: "dentist",
+    namePool: ["Dr. Pearl", "Dr. Molar", "Dr. Varnish", "Dr. Crown"],
+    salary: [780, 980],
+    traits: ["Gentle touch", "Laser focus", "Minty", "Cosmetic ace"],
+  },
+  {
+    role: "chef",
+    namePool: ["Chef Basil", "Chef Lumen", "Chef Pico", "Chef Dahlia"],
+    salary: [560, 760],
+    traits: ["Flambé", "Comfort food", "Nutritionist", "Speedy"],
+  },
+];
+
