+++ main.js
@@ -1,1225 +1,65 @@
-const GRID_WIDTH = 12;
-const GRID_HEIGHT = 8;
-const STARTING_CASH = 150000;
-const DAY_TICKS = 60;
-const CANVAS_CELL = 56;
-const AXIS_LETTERS = "ABCDEFGHIJKLMNOPQRSTUVWXYZ";
-const SAVE_VERSION = 1;
-const SAVE_SLOT_COUNT = 4;
-const SAVE_STORAGE_KEY_PREFIX = "pulse-point-hospital-save";
-const VIEW_MODE_STORAGE_KEY = "pulse-point-hospital-view-mode";
-const ISO_TILE_WIDTH = 64;
-const ISO_TILE_HEIGHT = 36;
-const ISO_WALL_HEIGHT = 28;
-const NEW_ROOM_HIGHLIGHT_TICKS = 36;
-const MIN_SHOWCASE_ZOOM = 0.65;
-const MAX_SHOWCASE_ZOOM = 1.8;
-const SHOWCASE_PAN_PADDING = 0.35;
-const AUDIO_SETTINGS_STORAGE_KEY = "pulse-point-hospital-audio-settings";
-const DEFAULT_AUDIO_SETTINGS = { music: true, sfx: true };
-const BACKGROUND_MUSIC_DURATION = 8;
+import {
+  GRID_WIDTH,
+  GRID_HEIGHT,
+  STARTING_CASH,
+  DAY_TICKS,
+  CANVAS_CELL,
+  AXIS_LETTERS,
+  SAVE_VERSION,
+  SAVE_SLOT_COUNT,
+  SAVE_STORAGE_KEY_PREFIX,
+  VIEW_MODE_STORAGE_KEY,
+  ISO_TILE_WIDTH,
+  ISO_TILE_HEIGHT,
+  ISO_WALL_HEIGHT,
+  NEW_ROOM_HIGHLIGHT_TICKS,
+  MIN_SHOWCASE_ZOOM,
+  MAX_SHOWCASE_ZOOM,
+  SHOWCASE_PAN_PADDING,
+  AUDIO_SETTINGS_STORAGE_KEY,
+  DEFAULT_AUDIO_SETTINGS,
+  BACKGROUND_MUSIC_DURATION,
+  CORRIDOR_Y,
+  ENTRANCE_X,
+  AGENT_PATIENT_SPEED,
+  AGENT_STAFF_SPEED,
+  TREATMENT_DURATION_TICKS,
+} from "./js/constants.js";
+import {
+  PROPERTY_PARCELS,
+  ROOM_SIZE_LIBRARY,
+  ROOM_INTERIOR_LIBRARY,
+  ROOM_MACHINE_LIBRARY,
+  ROOM_DECOR_LIBRARY,
+  roomCatalog,
+  baseSeverityMap,
+  defaultDesignProfile,
+} from "./js/data/roomData.js";
+import { staffCatalog } from "./js/data/staffData.js";
+import {
+  billingProfiles,
+  patientNames,
+  loanOffers,
+  researchProjects,
+  marketingCampaigns,
+  ailments,
+  emergencyCases,
+  objectives,
+  restoreObjectives,
+} from "./js/data/gameData.js";
+import {
+  deepClone,
+  clamp,
+  hashString,
+  shiftColor,
+  hexToRgb,
+  withAlpha,
+} from "./js/utils/helpers.js";
+
 const showcaseAnimation = {
   time: 0,
   pulse: 0,
   wave: 0,
-};
-const CORRIDOR_Y = GRID_HEIGHT - 0.25;
-const ENTRANCE_X = 1.5;
-const AGENT_PATIENT_SPEED = 0.18;
-const AGENT_STAFF_SPEED = 0.22;
-const TREATMENT_DURATION_TICKS = 18;
-
-const PROPERTY_PARCELS = [
-  {
-    id: "atrium",
-    name: "Founders Atrium",
-    cost: 0,
-    description: "Starter plaza with enough space for a GP hub and diagnostics.",
-    x: 0,
-    y: 0,
-    width: 6,
-    height: 4,
-  },
-  {
-    id: "research",
-    name: "Innovation Lot",
-    cost: 48000,
-    description: "Unlocks the north labs and plenty of room for specialty clinics.",
-    x: 6,
-    y: 0,
-    width: 6,
-    height: 4,
-  },
-  {
-    id: "gardens",
-    name: "Wellness Gardens",
-    cost: 36000,
-    description: "Adds calming grounds for wards, psychiatry, and guest services.",
-    x: 0,
-    y: 4,
-    width: 6,
-    height: 4,
-  },
-  {
-    id: "emergency",
-    name: "Emergency Concourse",
-    cost: 52000,
-    description: "Extends the trauma wing for emergency, surgery, and imaging suites.",
-    x: 6,
-    y: 4,
-    width: 6,
-    height: 4,
-  },
-];
-
-const ROOM_SIZE_LIBRARY = [
-  {
-    id: "compact",
-    label: "Compact (1x1)",
-    width: 1,
-    height: 1,
-    costMultiplier: 1,
-    upkeepMultiplier: 1,
-    severityBonus: 0,
-    environmentBonus: 0.2,
-    description: "Entry layout for a quick setup.",
-  },
-  {
-    id: "standard",
-    label: "Standard (2x1)",
-    width: 2,
-    height: 1,
-    costMultiplier: 1.35,
-    upkeepMultiplier: 1.2,
-    severityBonus: 0.5,
-    environmentBonus: 0.6,
-    description: "Adds an extra bay for machines and visitors.",
-  },
-  {
-    id: "deluxe",
-    label: "Deluxe (2x2)",
-    width: 2,
-    height: 2,
-    costMultiplier: 1.8,
-    upkeepMultiplier: 1.35,
-    severityBonus: 1,
-    environmentBonus: 1,
-    description: "Premium suite with ample space for advanced care.",
-  },
-];
-
-const ROOM_INTERIOR_LIBRARY = [
-  {
-    id: "modern",
-    label: "Modern Steel",
-    cost: 1200,
-    upkeep: 35,
-    palette: {
-      base: "#1f2937",
-      mid: "#334155",
-      accent: "#60a5fa",
-      decor: "#93c5fd",
-    },
-    modifiers: { efficiency: 2, morale: 1 },
-  },
-  {
-    id: "soothing",
-    label: "Soothing Pastels",
-    cost: 1500,
-    upkeep: 30,
-    palette: {
-      base: "#fdf2f8",
-      mid: "#fce7f3",
-      accent: "#f472b6",
-      decor: "#34d399",
-    },
-    modifiers: { welfare: 3, environment: 2 },
-  },
-  {
-    id: "vibrant",
-    label: "Vibrant Pop",
-    cost: 1000,
-    upkeep: 25,
-    palette: {
-      base: "#0f172a",
-      mid: "#1e293b",
-      accent: "#f97316",
-      decor: "#facc15",
-    },
-    modifiers: { reputation: 2, morale: 2 },
-  },
-  {
-    id: "nature",
-    label: "Nature Retreat",
-    cost: 1350,
-    upkeep: 28,
-    palette: {
-      base: "#f1f5f9",
-      mid: "#bbf7d0",
-      accent: "#22c55e",
-      decor: "#65a30d",
-    },
-    modifiers: { welfare: 2, morale: 1, environment: 3 },
-  },
-  {
-    id: "art-deco",
-    label: "Art Deco Luxe",
-    cost: 1650,
-    upkeep: 38,
-    palette: {
-      base: "#0f172a",
-      mid: "#1e293b",
-      accent: "#fbbf24",
-      decor: "#f472b6",
-    },
-    modifiers: { reputation: 3, morale: 2 },
-  },
-  {
-    id: "sterile",
-    label: "Sterile Quartz",
-    cost: 1400,
-    upkeep: 32,
-    palette: {
-      base: "#f8fafc",
-      mid: "#cbd5f5",
-      accent: "#38bdf8",
-      decor: "#94a3b8",
-    },
-    modifiers: { efficiency: 2, environment: 1 },
-  },
-  {
-    id: "sunset",
-    label: "Sunset Oasis",
-    cost: 1550,
-    upkeep: 34,
-    palette: {
-      base: "#fff7ed",
-      mid: "#fed7aa",
-      accent: "#fb7185",
-      decor: "#f97316",
-    },
-    modifiers: { welfare: 2, morale: 2 },
-  },
-  {
-    id: "zenith",
-    label: "Zenith Glasshouse",
-    cost: 1850,
-    upkeep: 38,
-    palette: {
-      base: "#0b1120",
-      mid: "#134e4a",
-      accent: "#2dd4bf",
-      decor: "#f59e0b",
-    },
-    modifiers: { environment: 3, reputation: 1, efficiency: 1 },
-  },
-  {
-    id: "heritage",
-    label: "Heritage Copper",
-    cost: 1700,
-    upkeep: 33,
-    palette: {
-      base: "#1c1917",
-      mid: "#b45309",
-      accent: "#f97316",
-      decor: "#fde68a",
-    },
-    modifiers: { morale: 2, reputation: 2 },
-  },
-  {
-    id: "aurora",
-    label: "Aurora Pulse",
-    cost: 1780,
-    upkeep: 37,
-    palette: {
-      base: "#111827",
-      mid: "#312e81",
-      accent: "#a855f7",
-      decor: "#38bdf8",
-    },
-    modifiers: { efficiency: 2, welfare: 1, reputation: 2 },
-  },
-];
-
-const ROOM_MACHINE_LIBRARY = [
-  {
-    id: "standard-kit",
-    label: "Standard Treatment Kit",
-    cost: 0,
-    upkeep: 0,
-    severityBoost: 0,
-    welfareBoost: 0,
-    description: "Baseline tools for routine care.",
-    color: "#cbd5f5",
-  },
-  {
-    id: "advanced-diagnostics",
-    label: "Advanced Diagnostics Rig",
-    cost: 4200,
-    upkeep: 80,
-    severityBoost: 1,
-    welfareBoost: 0.5,
-    description: "High-resolution scanners for tricky ailments.",
-    color: "#38bdf8",
-  },
-  {
-    id: "precision-suite",
-    label: "Precision Therapy Suite",
-    cost: 7600,
-    upkeep: 120,
-    severityBoost: 2,
-    welfareBoost: 1,
-    description: "Specialist pods that unlock elite treatments.",
-    color: "#f97316",
-  },
-  {
-    id: "holistic-support",
-    label: "Holistic Support Capsule",
-    cost: 3500,
-    upkeep: 55,
-    severityBoost: 0.3,
-    welfareBoost: 1.5,
-    description: "Comfort tech that keeps patients calm and cooperative.",
-    color: "#34d399",
-  },
-];
-
-const ROOM_DECOR_LIBRARY = [
-  {
-    id: "calming-plants",
-    label: "Calming Planters",
-    cost: 800,
-    upkeep: 10,
-    environment: 2,
-    welfare: 1,
-    morale: 1,
-    reputation: 0,
-    description: "Greenery that softens clinical edges.",
-    color: "#22c55e",
-    isPlant: true,
-  },
-  {
-    id: "art-installation",
-    label: "Art Installation",
-    cost: 1100,
-    upkeep: 14,
-    environment: 1,
-    welfare: 1,
-    morale: 0,
-    reputation: 1,
-    description: "Gallery-worthy pieces for a prestige boost.",
-    color: "#a855f7",
-  },
-  {
-    id: "luxury-seating",
-    label: "Luxury Seating",
-    cost: 1500,
-    upkeep: 18,
-    environment: 1,
-    welfare: 2,
-    morale: 2,
-    reputation: 1,
-    description: "Plush lounges that keep visitors smiling.",
-    color: "#facc15",
-  },
-  {
-    id: "healing-sculpture",
-    label: "Healing Sculpture",
-    cost: 950,
-    upkeep: 12,
-    environment: 1,
-    welfare: 1,
-    morale: 1,
-    reputation: 1,
-    description: "Centerpiece art celebrating recovery stories.",
-    color: "#0ea5e9",
-  },
-  {
-    id: "snack-vendor",
-    label: "Snack Machine",
-    cost: 650,
-    upkeep: 8,
-    environment: 0.2,
-    welfare: 1,
-    morale: 1,
-    reputation: 0,
-    description: "Dispenses quick bites that cheer up waiting patients (and create a little litter).",
-    color: "#fb7185",
-    amenityType: "snack",
-    amenityLabel: "Snacks",
-  },
-  {
-    id: "atm-kiosk",
-    label: "ATM Kiosk",
-    cost: 950,
-    upkeep: 9,
-    environment: 0.5,
-    welfare: 0,
-    morale: 0,
-    reputation: 1,
-    description: "Cash-point services that earn small fees and reassure guests.",
-    color: "#0ea5e9",
-    amenityType: "atm",
-    amenityLabel: "ATM",
-  },
-  {
-    id: "hydration-station",
-    label: "Hydration Station",
-    cost: 700,
-    upkeep: 7,
-    environment: 1,
-    welfare: 1,
-    morale: 0,
-    reputation: 0,
-    description: "Water coolers and cups that keep everyone refreshed.",
-    color: "#60a5fa",
-    amenityType: "hydration",
-    amenityLabel: "Hydration",
-  },
-  {
-    id: "hanging-garden",
-    label: "Hanging Garden",
-    cost: 900,
-    upkeep: 11,
-    environment: 2,
-    welfare: 1,
-    morale: 1,
-    reputation: 0,
-    description: "Vertical planters that need routine watering but elevate the ambience.",
-    color: "#4ade80",
-    isPlant: true,
-  },
-];
-
-const roomCatalog = [
-  {
-    id: "reception",
-    name: "Reception",
-    cost: 4500,
-    upkeep: 150,
-    roleRequired: ["assistant"],
-    reputation: 1,
-    description: "Welcomes new arrivals and keeps the queue orderly.",
-  },
-  {
-    id: "triage",
-    name: "Triage Desk",
-    cost: 5200,
-    upkeep: 180,
-    roleRequired: ["assistant", "nurse"],
-    reputation: 2,
-    description: "Sorts arrivals and smooths the waiting line before treatment.",
-  },
-  {
-    id: "gp",
-    name: "Doctor's Office",
-    cost: 6000,
-    upkeep: 200,
-    roleRequired: ["doctor"],
-    reputation: 2,
-    description: "Diagnoses patient ailments and sets their treatment plan.",
-  },
-  {
-    id: "psychiatry",
-    name: "Psychiatry",
-    cost: 14000,
-    upkeep: 360,
-    roleRequired: ["doctor"],
-    reputation: 3,
-    description: "Provides therapy for mental health disorders and stress ailments.",
-  },
-  {
-    id: "pediatrics",
-    name: "Pediatrics",
-    cost: 13000,
-    upkeep: 340,
-    roleRequired: ["doctor", "nurse"],
-    reputation: 3,
-    description: "Bright ward for children with extra play space and comfort.",
-  },
-  {
-    id: "maternity",
-    name: "Maternity Suite",
-    cost: 17000,
-    upkeep: 420,
-    roleRequired: ["doctor", "midwife"],
-    reputation: 4,
-    description: "Delivers new arrivals with calming decor and specialised care.",
-  },
-  {
-    id: "pharmacy",
-    name: "Pharmacy",
-    cost: 8000,
-    upkeep: 250,
-    roleRequired: ["nurse"],
-    reputation: 3,
-    description: "Dispenses medicine for low severity conditions.",
-  },
-  {
-    id: "ward",
-    name: "Ward",
-    cost: 12000,
-    upkeep: 360,
-    roleRequired: ["doctor", "nurse"],
-    reputation: 4,
-    description: "Treats moderate conditions requiring overnight care.",
-  },
-  {
-    id: "diagnostics",
-    name: "Diagnostics Lab",
-    cost: 15000,
-    upkeep: 420,
-    roleRequired: ["doctor"],
-    reputation: 4,
-    description: "Runs advanced tests to speed up treatment chains.",
-  },
-  {
-    id: "cardiology",
-    name: "Cardiology",
-    cost: 20000,
-    upkeep: 520,
-    roleRequired: ["doctor", "nurse"],
-    reputation: 5,
-    description: "Monitors heart health and treats rhythm-related illnesses.",
-  },
-  {
-    id: "orthopedics",
-    name: "Orthopedics",
-    cost: 19000,
-    upkeep: 480,
-    roleRequired: ["doctor", "therapist"],
-    reputation: 4,
-    description: "Fixes broken bones and mobility issues with rehab stations.",
-  },
-  {
-    id: "neurology",
-    name: "Neurology",
-    cost: 21000,
-    upkeep: 540,
-    roleRequired: ["doctor", "technician"],
-    reputation: 5,
-    description: "Maps brain activity with scanners and quiet monitoring pods.",
-  },
-  {
-    id: "dermatology",
-    name: "Dermatology",
-    cost: 10000,
-    upkeep: 260,
-    roleRequired: ["doctor"],
-    reputation: 2,
-    description: "Treats skin conditions with UV bays and soothing creams.",
-  },
-  {
-    id: "gastro",
-    name: "Gastroenterology",
-    cost: 16000,
-    upkeep: 400,
-    roleRequired: ["doctor", "nurse"],
-    reputation: 4,
-    description: "Investigates digestive disorders with scopes and recovery chairs.",
-  },
-  {
-    id: "dialysis",
-    name: "Dialysis",
-    cost: 15000,
-    upkeep: 420,
-    roleRequired: ["nurse", "technician"],
-    reputation: 4,
-    description: "Filters blood with high-tech machines for chronic patients.",
-  },
-  {
-    id: "physiotherapy",
-    name: "Physiotherapy",
-    cost: 12500,
-    upkeep: 320,
-    roleRequired: ["nurse"],
-    reputation: 3,
-    description: "Rehabilitates injuries and mobility issues with guided exercise.",
-  },
-  {
-    id: "oncology",
-    name: "Oncology",
-    cost: 22000,
-    upkeep: 560,
-    roleRequired: ["doctor", "technician"],
-    reputation: 5,
-    description: "Provides chemo bays and support zones for cancer patients.",
-  },
-  {
-    id: "burn",
-    name: "Burn Unit",
-    cost: 23000,
-    upkeep: 600,
-    roleRequired: ["doctor", "nurse", "therapist"],
-    reputation: 5,
-    description: "Cares for severe burns with hydro tanks and pain therapy booths.",
-  },
-  {
-    id: "dental",
-    name: "Dental Surgery",
-    cost: 9000,
-    upkeep: 260,
-    roleRequired: ["dentist", "assistant"],
-    reputation: 3,
-    description: "Restores smiles with chrome chairs and high-powered polishers.",
-  },
-  {
-    id: "ophthalmology",
-    name: "Ophthalmology",
-    cost: 18000,
-    upkeep: 520,
-    roleRequired: ["doctor", "technician"],
-    reputation: 4,
-    description: "Performs delicate eye work with laser rigs and dark testing rooms.",
-  },
-  {
-    id: "surgery",
-    name: "Operating Theatre",
-    cost: 26000,
-    upkeep: 900,
-    roleRequired: ["surgeon", "nurse"],
-    reputation: 6,
-    description: "Performs complex procedures for severe conditions.",
-  },
-  {
-    id: "icu",
-    name: "Intensive Care Unit",
-    cost: 24000,
-    upkeep: 820,
-    roleRequired: ["doctor", "nurse"],
-    reputation: 5,
-    description: "Stabilises critical patients and boosts emergency reputation.",
-  },
-  {
-    id: "emergency",
-    name: "Emergency Department",
-    cost: 24000,
-    upkeep: 780,
-    roleRequired: ["doctor", "nurse", "security"],
-    reputation: 6,
-    description: "Handles urgent arrivals with crash bays and trauma carts.",
-  },
-  {
-    id: "research",
-    name: "Research Lab",
-    cost: 18000,
-    upkeep: 550,
-    roleRequired: ["researcher"],
-    reputation: 5,
-    description: "Unlocks new cures and room upgrades.",
-  },
-  {
-    id: "cafeteria",
-    name: "Wellness Cafeteria",
-    cost: 9000,
-    upkeep: 260,
-    roleRequired: ["assistant"],
-    reputation: 2,
-    description: "Keeps patients waiting happy and lifts staff morale.",
-  },
-  {
-    id: "gourmet",
-    name: "Gourmet Kitchen",
-    cost: 11000,
-    upkeep: 320,
-    roleRequired: ["chef", "assistant"],
-    reputation: 3,
-    description: "Serves signature meals to energise staff and pamper VIPs.",
-  },
-  {
-    id: "giftshop",
-    name: "Gift Shop",
-    cost: 6500,
-    upkeep: 160,
-    roleRequired: ["assistant"],
-    reputation: 2,
-    description: "Sells plush mascots and boosts visitor satisfaction.",
-  },
-  {
-    id: "clowncare",
-    name: "Clown Care Ward",
-    cost: 9000,
-    upkeep: 260,
-    roleRequired: ["entertainer"],
-    reputation: 3,
-    description: "Inspired by Theme Hospital antics to keep spirits high.",
-  },
-  {
-    id: "holotheatre",
-    name: "Holotheatre",
-    cost: 12500,
-    upkeep: 320,
-    roleRequired: ["entertainer", "assistant"],
-    reputation: 3,
-    description: "Projects Biing-style cabaret holograms minus the risqué bits.",
-  },
-  {
-    id: "zen",
-    name: "Zen Garden",
-    cost: 8000,
-    upkeep: 220,
-    roleRequired: ["therapist"],
-    reputation: 2,
-    description: "Restorative retreat with koi ponds for stressed staff.",
-  },
-  {
-    id: "securityoffice",
-    name: "Security Office",
-    cost: 7000,
-    upkeep: 200,
-    roleRequired: ["security"],
-    reputation: 2,
-    description: "Keeps troublemakers in check and smooths emergency response.",
-  },
-  {
-    id: "billing",
-    name: "Billing Office",
-    cost: 10500,
-    upkeep: 300,
-    roleRequired: ["accountant"],
-    reputation: 3,
-    description: "Optimises invoices for better payouts and financial tracking.",
-  },
-  {
-    id: "marketing",
-    name: "Public Relations Suite",
-    cost: 9500,
-    upkeep: 220,
-    roleRequired: ["marketer"],
-    reputation: 2,
-    description: "Launch campaigns to attract lucrative patient groups.",
-  },
-  {
-    id: "staffroom",
-    name: "Staff Lounge",
-    cost: 7000,
-    upkeep: 160,
-    roleRequired: [],
-    reputation: 1,
-    description: "Boosts staff morale and reduces burnout.",
-  },
-];
-
-const baseSeverityMap = {
-  reception: 0,
-  triage: 1,
-  gp: 2,
-  psychiatry: 2,
-  pediatrics: 2,
-  maternity: 3,
-  pharmacy: 1,
-  ward: 2,
-  diagnostics: 2,
-  cardiology: 3,
-  orthopedics: 3,
-  neurology: 3,
-  dermatology: 2,
-  gastro: 3,
-  dialysis: 3,
-  physiotherapy: 2,
-  oncology: 4,
-  burn: 4,
-  dental: 2,
-  ophthalmology: 3,
-  surgery: 4,
-  icu: 4,
-  emergency: 3,
-  cafeteria: 0,
-  gourmet: 0,
-  giftshop: 0,
-  billing: 0,
-  marketing: 0,
-  clowncare: 1,
-  holotheatre: 1,
-  zen: 1,
-  research: 0,
-  securityoffice: 0,
-  staffroom: 0,
-  radiology: 3,
-};
-
-const defaultDesignProfile = {
-  sizes: ROOM_SIZE_LIBRARY.map((option) => option.id),
-  interiors: ROOM_INTERIOR_LIBRARY.map((theme) => theme.id),
-  machines: ROOM_MACHINE_LIBRARY.map((machine) => machine.id),
-  decor: ROOM_DECOR_LIBRARY.map((decor) => decor.id),
-  defaultMachines: ["standard-kit"],
-};
-
-roomCatalog.forEach((room) => {
-  room.baseSeverity = baseSeverityMap[room.id] ?? 1;
-  room.designProfile = {
-    sizes: [...defaultDesignProfile.sizes],
-    interiors: [...defaultDesignProfile.interiors],
-    machines:
-      (baseSeverityMap[room.id] ?? 1) > 0 ? [...defaultDesignProfile.machines] : ["standard-kit"],
-    decor: [...defaultDesignProfile.decor],
-    defaultMachines: [...defaultDesignProfile.defaultMachines],
-  };
-  if ((baseSeverityMap[room.id] ?? 1) <= 0) {
-    room.designProfile.machines = ["standard-kit"];
-  }
-});
-
-const staffCatalog = [
-  {
-    role: "doctor",
-    namePool: ["Dr. Nova Halley", "Dr. Elias Quill", "Dr. Miri Vega", "Dr. Jun Pierce"],
-    salary: [900, 1100],
-    traits: ["Bedside manner", "Diagnostic genius", "Fast learner", "Research enthusiast"],
-  },
-  {
-    role: "nurse",
-    namePool: ["Nurse Aiko", "Nurse Lucien", "Nurse Piper", "Nurse Saffron"],
-    salary: [650, 850],
-    traits: ["Efficient", "Empathetic", "Inventory hawk", "Inspiring"],
-  },
-  {
-    role: "assistant",
-    namePool: ["Maya Front", "Theo Ledger", "Rae Burst", "Ivy Kios"],
-    salary: [450, 600],
-    traits: ["Charismatic", "Queue master", "Patient whisperer", "Energetic"],
-  },
-  {
-    role: "janitor",
-    namePool: ["Vik Sweep", "Chaz Buff", "Vera Shine", "Kato Mop"],
-    salary: [380, 520],
-    traits: ["Meticulous", "Fast cleaner", "Mechanic", "Green thumb"],
-  },
-  {
-    role: "facility",
-    namePool: ["Mina Grove", "Odin Wells", "Clara Bloom", "Zed Rinse"],
-    salary: [420, 600],
-    traits: ["Plant whisperer", "Spill scout", "Sparkling tidy", "Logistics ace"],
-  },
-  {
-    role: "researcher",
-    namePool: ["Dr. Lyra Code", "Analyst Vex", "Dr. Reva Coil", "Milo Synth"],
-    salary: [700, 950],
-    traits: ["Innovator", "Grant magnet", "Focused", "Collaborative"],
-  },
-  {
-    role: "marketer",
-    namePool: ["Ada Pitch", "Rowan Buzz", "Nix Viral", "Tara Prism"],
-    salary: [500, 680],
-    traits: ["Storyteller", "Analytical", "Hype builder", "Networked"],
-  },
-  {
-    role: "surgeon",
-    namePool: ["Dr. Mira Scalpel", "Dr. Orion Bloom", "Dr. Zia Voss", "Dr. Helio Mars"],
-    salary: [1200, 1500],
-    traits: ["Calm hands", "Rapid recovery", "Team lead", "Minimal scarring"],
-  },
-  {
-    role: "technician",
-    namePool: ["Ray Pulse", "Iris Beam", "Nico Magnet", "Kara Scan"],
-    salary: [580, 760],
-    traits: ["Precision", "Calibrates fast", "Empathic", "Data driven"],
-  },
-  {
-    role: "accountant",
-    namePool: ["Pia Ledger", "Miles Tally", "June Audit", "Cai Balance"],
-    salary: [520, 720],
-    traits: ["Eagle eye", "Negotiator", "Budget hawk", "Calm"],
-  },
-  {
-    role: "midwife",
-    namePool: ["Mara Bloom", "Leah Cradle", "Ina Dove", "Ruth Lulla"],
-    salary: [700, 900],
-    traits: ["Gentle", "Quick hands", "Reassuring", "Songbird"],
-  },
-  {
-    role: "therapist",
-    namePool: ["Galen Flex", "Noor Mend", "Ravi Align", "Elara Sooth"],
-    salary: [680, 880],
-    traits: ["Motivator", "Massage pro", "Mindful", "Stretch guru"],
-  },
-  {
-    role: "entertainer",
-    namePool: ["Bobo Spark", "Luna Gag", "Pip Confetti", "Dex Pratfall"],
-    salary: [520, 720],
-    traits: ["Slapstick", "Improv", "Balloon artist", "Banters"],
-  },
-  {
-    role: "security",
-    namePool: ["Axel Guard", "Mira Watch", "Jett Shield", "Tova Brace"],
-    salary: [550, 750],
-    traits: ["Alert", "Crowd control", "Cool head", "Swift"],
-  },
-  {
-    role: "dentist",
-    namePool: ["Dr. Pearl", "Dr. Molar", "Dr. Varnish", "Dr. Crown"],
-    salary: [780, 980],
-    traits: ["Gentle touch", "Laser focus", "Minty", "Cosmetic ace"],
-  },
-  {
-    role: "chef",
-    namePool: ["Chef Basil", "Chef Lumen", "Chef Pico", "Chef Dahlia"],
-    salary: [560, 760],
-    traits: ["Flambé", "Comfort food", "Nutritionist", "Speedy"],
-  },
-];
-
-const billingProfiles = [
-  {
-    id: "insured-basic",
-    label: "Insurance - Basic",
-    modifier: 0.9,
-    coverage: 0.75,
-    copay: 125,
-    method: "insurance",
-    color: "#38bdf8",
-  },
-  {
-    id: "insured-premium",
-    label: "Insurance - Premium",
-    modifier: 1.05,
-    coverage: 0.9,
-    copay: 80,
-    method: "insurance",
-    color: "#0ea5e9",
-  },
-  {
-    id: "self-credit",
-    label: "Self-Pay (Credit)",
-    modifier: 1.1,
-    coverage: 0,
-    processingFee: 0.025,
-    method: "credit",
-    color: "#f97316",
-  },
-  {
-    id: "self-cash",
-    label: "Self-Pay (Cash)",
-    modifier: 0.95,
-    coverage: 0,
-    discount: 0.05,
-    method: "cash",
-    color: "#facc15",
-  },
-  {
-    id: "installment",
-    label: "Installment Plan",
-    modifier: 1,
-    coverage: 0,
-    depositRate: 0.2,
-    term: 10,
-    method: "installment",
-    color: "#a855f7",
-  },
-  {
-    id: "charity",
-    label: "Charity Care",
-    modifier: 0.7,
-    coverage: 0.5,
-    copay: 0,
-    method: "insurance",
-    color: "#c084fc",
-    discount: 0.2,
-  },
-];
-
-const patientNames = [
-  "Alex",
-  "Charlie",
-  "Dakota",
-  "Jules",
-  "Robin",
-  "Sasha",
-  "Emerson",
-  "Hayden",
-  "Nova",
-  "Ezra",
-  "Finley",
-  "Kai",
-  "River",
-  "Skye",
-  "Rowan",
-  "Phoenix",
-];
-
-const loanOffers = [
-  {
-    id: "starter",
-    name: "Starter Expansion Loan",
-    amount: 25000,
-    term: 8,
-    interestRate: 0.12,
-    desc: "Short-term cash for early renovations. Paid back over eight days.",
-  },
-  {
-    id: "growth",
-    name: "Growth Capital Loan",
-    amount: 50000,
-    term: 14,
-    interestRate: 0.18,
-    desc: "Fuel new departments with a longer repayment window and higher interest.",
-  },
-  {
-    id: "infrastructure",
-    name: "Infrastructure Bond",
-    amount: 90000,
-    term: 20,
-    interestRate: 0.25,
-    desc: "Big investment for state-of-the-art wings — mind the sizable repayments.",
-  },
-];
-
-const researchProjects = [
-  {
-    id: "imaging",
-    name: "Advanced Imaging",
-    desc: "Unlocks a Radiology ward that doubles diagnosis accuracy.",
-    cost: 3000,
-    progress: 0,
-    unlocked: false,
-  },
-  {
-    id: "biofilter",
-    name: "Bio-filtered Air",
-    desc: "Boosts cleanliness and reduces infection risk hospital-wide.",
-    cost: 2500,
-    progress: 0,
-    unlocked: false,
-  },
-  {
-    id: "wellness",
-    name: "Holistic Wellness",
-    desc: "Adds a spa upgrade to the Staff Lounge for morale.",
-    cost: 2000,
-    progress: 0,
-    unlocked: false,
-  },
-];
-
-const marketingCampaigns = [
-  {
-    id: "prime",
-    name: "Prime Time Media Blast",
-    desc: "Boosts daily patient flow by 30% for three days.",
-    cost: 4000,
-    duration: 3,
-  },
-  {
-    id: "investors",
-    name: "Investor Showcase",
-    desc: "Gain an instant ¤6000 infusion but scrutiny lowers reputation if morale drops.",
-    cost: 2500,
-    duration: 1,
-  },
-  {
-    id: "community",
-    name: "Community Health Fair",
-    desc: "Raises reputation by 5 and attracts low-income patients (lower fees).",
-    cost: 1800,
-    duration: 2,
-  },
-];
-
-const ailments = [
-  { id: "coldsnap", name: "Cold Snap", severity: 1, room: "gp", payout: 850 },
-  { id: "sugarcrash", name: "Sugar Crash", severity: 1, room: "pharmacy", payout: 900 },
-  { id: "flatline", name: "Flatline Fever", severity: 2, room: "ward", payout: 1600 },
-  { id: "fractured", name: "Fractured Ego", severity: 2, room: "gp", payout: 1200 },
-  { id: "burnout", name: "Starlight Burnout", severity: 2, room: "psychiatry", payout: 1800 },
-  { id: "panic", name: "Cosmic Panic", severity: 3, room: "psychiatry", payout: 2600 },
-  { id: "phantom", name: "Phantom Limelight", severity: 3, room: "diagnostics", payout: 2200 },
-  { id: "appendix", name: "Appendix Boom", severity: 3, room: "surgery", payout: 3200 },
-  { id: "cardio", name: "Rhythm Crash", severity: 3, room: "cardiology", payout: 3400 },
-  { id: "sprain", name: "Nebula Sprain", severity: 2, room: "physiotherapy", payout: 1500 },
-  { id: "stiff", name: "Stellar Stiffness", severity: 1, room: "physiotherapy", payout: 1100 },
-  { id: "hypergly", name: "Hyper Glow", severity: 2, room: "cafeteria", payout: 950 },
-  { id: "papertrail", name: "Paper Trail Phobia", severity: 1, room: "billing", payout: 700 },
-  { id: "meteormeasles", name: "Meteor Measles", severity: 2, room: "pediatrics", payout: 1700 },
-  { id: "toybox", name: "Toy Box Fever", severity: 1, room: "pediatrics", payout: 1200 },
-  { id: "lunarcramps", name: "Lunar Cramps", severity: 3, room: "maternity", payout: 3000 },
-  { id: "spacecolic", name: "Space Colic", severity: 2, room: "maternity", payout: 2100 },
-  { id: "asteroidache", name: "Asteroid Ache", severity: 2, room: "orthopedics", payout: 2300 },
-  { id: "gravitygrind", name: "Gravity Grind", severity: 2, room: "orthopedics", payout: 2400 },
-  { id: "neuroflash", name: "Neuro Flash", severity: 3, room: "neurology", payout: 3200 },
-  { id: "synapse", name: "Synapse Static", severity: 4, room: "neurology", payout: 3600 },
-  { id: "solarrash", name: "Solar Rash", severity: 1, room: "dermatology", payout: 950 },
-  { id: "moonburn", name: "Moon Burn", severity: 2, room: "dermatology", payout: 1500 },
-  { id: "warpbelly", name: "Warp Belly", severity: 3, room: "gastro", payout: 2800 },
-  { id: "stomavortex", name: "Stoma Vortex", severity: 2, room: "gastro", payout: 2100 },
-  { id: "nebularinse", name: "Nebula Rinse", severity: 3, room: "dialysis", payout: 2700 },
-  { id: "plasmabuild", name: "Plasma Build-Up", severity: 4, room: "dialysis", payout: 3100 },
-  { id: "quantumcells", name: "Quantum Cells", severity: 4, room: "oncology", payout: 3800 },
-  { id: "supernova", name: "Supernova Stress", severity: 3, room: "oncology", payout: 3400 },
-  { id: "emberveil", name: "Ember Veil", severity: 4, room: "burn", payout: 3600 },
-  { id: "flareburst", name: "Flare Burst", severity: 3, room: "burn", payout: 3100 },
-  { id: "cosmiccavity", name: "Cosmic Cavity", severity: 1, room: "dental", payout: 1200 },
-  { id: "jawsaw", name: "Jaw Saw", severity: 2, room: "dental", payout: 2000 },
-  { id: "galaxyglare", name: "Galaxy Glare", severity: 2, room: "ophthalmology", payout: 2200 },
-  { id: "nebulacurtain", name: "Nebula Curtain", severity: 3, room: "ophthalmology", payout: 2800 },
-  { id: "laughitis", name: "Laugh-itis", severity: 1, room: "clowncare", payout: 950 },
-  { id: "sadzap", name: "Sad Zap", severity: 2, room: "holotheatre", payout: 1500 },
-  { id: "zenless", name: "Zenless Zing", severity: 1, room: "zen", payout: 1000 },
-];
-
-ailments.forEach((ailment) => {
-  const base = ailment.payout;
-  const severityBoost = 1 + (ailment.severity ?? 1) * 0.12;
-  const baseline = base * 1.25 * severityBoost;
-  ailment.charge = Math.round(baseline / 10) * 10;
-});
-
-const emergencyCases = [
-  { id: "superbug", name: "Superbug Outbreak", severity: 4, room: "icu", payout: 4200 },
-  { id: "heartquake", name: "Heartquake", severity: 4, room: "surgery", payout: 4600 },
-  { id: "voidfracture", name: "Void Fracture", severity: 4, room: "radiology", payout: 3800 },
-  { id: "soulstorm", name: "Soul Storm", severity: 4, room: "psychiatry", payout: 3600 },
-  { id: "cardioblast", name: "Cardio Blast", severity: 4, room: "cardiology", payout: 4400 },
-  { id: "starlabor", name: "Starfall Labour", severity: 4, room: "maternity", payout: 4200 },
-  { id: "brainflare", name: "Brain Flare", severity: 5, room: "neurology", payout: 4700 },
-  { id: "toxicwave", name: "Toxic Wave", severity: 4, room: "gastro", payout: 4100 },
-  { id: "emberstorm", name: "Ember Storm", severity: 5, room: "burn", payout: 4800 },
-  { id: "voidmilk", name: "Void Milk Recall", severity: 4, room: "pediatrics", payout: 3900 },
-  { id: "opticblast", name: "Optic Blast", severity: 4, room: "ophthalmology", payout: 4300 },
-  { id: "dialysisurge", name: "Dialysis Surge", severity: 4, room: "dialysis", payout: 4000 },
-];
-
-emergencyCases.forEach((emergency) => {
-  const base = emergency.payout;
-  const severityBoost = 1 + (emergency.severity ?? 3) * 0.18;
-  const baseline = base * 1.3 * severityBoost;
-  emergency.charge = Math.round(baseline / 10) * 10;
-});
-
-const objectives = [
-  {
-    id: "build-reception",
-    text: "Build a Reception to guide arrivals.",
-    condition: (state) => state.rooms.some((r) => r.type === "reception"),
-    completed: false,
-  },
-  {
-    id: "hire-doctor",
-    text: "Hire at least two doctors.",
-    condition: (state) => state.staff.filter((s) => s.role === "doctor").length >= 2,
-    completed: false,
-  },
-  {
-    id: "treat-ten",
-    text: "Treat 10 patients.",
-    condition: (state) => state.stats.patientsTreated >= 10,
-    completed: false,
-  },
-  {
-    id: "reach-rep",
-    text: "Reach reputation 75.",
-    condition: (state) => state.stats.reputation >= 75,
-    completed: false,
-  },
-];
-
-const deepClone = (input) => JSON.parse(JSON.stringify(input));
-
-const restoreObjectives = (progress = []) => {
-  const base = objectives.map((objective) => ({ ...objective }));
-  if (!Array.isArray(progress)) {
-    return base;
-  }
-  progress.forEach((saved) => {
-    const target = base.find((objective) => objective.id === saved.id);
-    if (target) {
-      target.completed = Boolean(saved.completed);
-    }
-  });
-  return base;
-};
-
-const clamp = (value, min, max) => Math.max(min, Math.min(max, value));
-
-const hashString = (value) => {
-  if (!value) return 0;
-  let hash = 0;
-  for (let i = 0; i < value.length; i += 1) {
-    hash = (hash << 5) - hash + value.charCodeAt(i);
-    hash |= 0;
-  }
-  return Math.abs(hash);
-};
-
-const normalizeHex = (hex) => {
-  if (!hex) return "#000000";
-  let value = hex.trim();
-  if (value.startsWith("rgb")) {
-    const parts = value
-      .replace(/rgba?\(/, "")
-      .replace(/\)/, "")
-      .split(",")
-      .map((part) => Number(part.trim()))
-      .filter((part, index) => index < 3 && Number.isFinite(part));
-    if (parts.length === 3) {
-      const [r, g, b] = parts.map((channel) => clamp(Math.round(channel), 0, 255));
-      return `#${((1 << 24) + (r << 16) + (g << 8) + b).toString(16).slice(1)}`;
-    }
-    return "#000000";
-  }
-  if (!value.startsWith("#")) {
-    value = `#${value}`;
-  }
-  if (value.length === 4) {
-    value = `#${value[1]}${value[1]}${value[2]}${value[2]}${value[3]}${value[3]}`;
-  }
-  if (value.length < 7) {
-    value = `${value}${"0".repeat(7 - value.length)}`;
-  }
-  return value.slice(0, 7).toLowerCase();
-};
-
-const shiftColor = (hex, percent = 0) => {
-  const value = normalizeHex(hex);
-  const amount = clamp(percent, -1, 1);
-  const num = parseInt(value.slice(1), 16);
-  let r = (num >> 16) & 0xff;
-  let g = (num >> 8) & 0xff;
-  let b = num & 0xff;
-  if (amount >= 0) {
-    r = Math.round(r + (255 - r) * amount);
-    g = Math.round(g + (255 - g) * amount);
-    b = Math.round(b + (255 - b) * amount);
-  } else {
-    const factor = 1 + amount;
-    r = Math.round(r * factor);
-    g = Math.round(g * factor);
-    b = Math.round(b * factor);
-  }
-  r = clamp(r, 0, 255);
-  g = clamp(g, 0, 255);
-  b = clamp(b, 0, 255);
-  return `#${((1 << 24) + (r << 16) + (g << 8) + b).toString(16).slice(1)}`;
-};
-
-const hexToRgb = (hex) => {
-  const value = normalizeHex(hex);
-  const num = parseInt(value.slice(1), 16);
-  return {
-    r: (num >> 16) & 0xff,
-    g: (num >> 8) & 0xff,
-    b: num & 0xff,
-  };
-};
-
-const withAlpha = (hex, alpha) => {
-  const { r, g, b } = hexToRgb(hex);
-  return `rgba(${r}, ${g}, ${b}, ${alpha})`;
-};
-
-const drawRoundedRect = (ctx, x, y, width, height, radius) => {
-  const r = Math.min(radius, width / 2, height / 2);
-  ctx.beginPath();
-  ctx.moveTo(x + r, y);
-  ctx.lineTo(x + width - r, y);
-  ctx.quadraticCurveTo(x + width, y, x + width, y + r);
-  ctx.lineTo(x + width, y + height - r);
-  ctx.quadraticCurveTo(x + width, y + height, x + width - r, y + height);
-  ctx.lineTo(x + r, y + height);
-  ctx.quadraticCurveTo(x, y + height, x, y + height - r);
-  ctx.lineTo(x, y + r);
-  ctx.quadraticCurveTo(x, y, x + r, y);
-  ctx.closePath();
 };
 
 const storage = (() => {
